requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic-settings==2.2.1